requirements/base.txt
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
django-environ==0.4.5
django==2.2.9
djangorestframework-jwt==1.11.0
djangorestframework==3.11.0
idna==2.8
jinja2==2.11.1
markupsafe==1.1.1
pyjwt==1.7.1
pytz==2019.3
requests==2.22.0
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.25.8
